Abstract API queries and types into a folder (DESC)
All the React Query hooks are now more easily re-usable in a shared lib/api folder, as well as all the types and common helpers. The rest of the frontend is now updated to import the queries from the new folders
