+++ js/src/app/_component/MiniLeaderboard.tsx
@@ -1,8 +1,8 @@
 import MiniLeaderboardSkeleton from "@/app/_component/skeletons/MiniLeaderboardSkeleton";
 import LeaderboardMetadata from "@/app/leaderboard/_components/LeaderboardMetadata/LeaderboardMetadata";
-import { useFullLeaderboardEntriesQuery } from "@/app/leaderboard/hooks";
 import LeaderboardCard from "@/components/ui/LeaderboardCard";
 import Toast from "@/components/ui/toast/Toast";
+import { useCurrentLeaderboardUsersQuery } from "@/lib/api/queries/leaderboard";
 import { theme } from "@/lib/theme";
 import { Table, Text, Tooltip } from "@mantine/core";
 import { IconCircleCheckFilled } from "@tabler/icons-react";
@@ -10,7 +10,7 @@ import { FaDiscord } from "react-icons/fa";
 import { SiLeetcode } from "react-icons/si";
 
 export default function MiniLeaderboardDesktop() {
-  const { data, status } = useFullLeaderboardEntriesQuery({ pageSize: 5 });
+  const { data, status } = useCurrentLeaderboardUsersQuery({ pageSize: 5 });
 
   if (status === "pending") {
     return <MiniLeaderboardSkeleton />;
