+++ js/src/app/dashboard/Dashboard.page.tsx
@@ -3,11 +3,11 @@ import OnboardingLeetcodeUser from "@/app/dashboard/_components/OnboardingLeetco
 import ProblemOfTheDay from "@/app/dashboard/_components/ProblemOfTheDay/ProblemOfTheDay";
 import RecentSubmissions from "@/app/dashboard/_components/RecentSubmissions/RecentSubmissions";
 import RefreshSubmissions from "@/app/dashboard/_components/RefreshSubmissions";
-import { useAuthQuery } from "@/app/login/hooks";
 import { Footer } from "@/components/ui/footer/Footer";
 import Header from "@/components/ui/header/Header";
 import Toast from "@/components/ui/toast/Toast";
 import ToastWithRedirect from "@/components/ui/toast/ToastWithRedirect";
+import { useAuthQuery } from "@/lib/api/queries/auth";
 import { Center, Flex, Loader, Title } from "@mantine/core";
 import { useMediaQuery } from "@mantine/hooks";
