+++ js/src/app/onboarding/Onboarding.page.tsx
@@ -1,9 +1,9 @@
-import { useAuthQuery } from "@/app/login/hooks";
 import UsernameForm from "@/app/onboarding/_components/UsernameForm";
 import { Footer } from "@/components/ui/footer/Footer";
 import Header from "@/components/ui/header/Header";
 import Toast from "@/components/ui/toast/Toast";
 import ToastWithRedirect from "@/components/ui/toast/ToastWithRedirect";
+import { useAuthQuery } from "@/lib/api/queries/auth";
 import { Loader } from "@mantine/core";
 
 export default function Onboarding() {
