+++ js/src/lib/api/queries/leaderboard/index.ts
@@ -1,0 +1,166 @@
+import { ApiResponse } from "@/lib/api/common/apiResponse";
+import { Page } from "@/lib/api/common/page";
+import { Leaderboard } from "@/lib/api/types/leaderboard";
+import { User, UserWithScore } from "@/lib/api/types/user";
+import { useURLState } from "@/lib/hooks/useUrlState";
+import {
+  keepPreviousData,
+  useMutation,
+  useQuery,
+  useQueryClient,
+} from "@tanstack/react-query";
+import { useCallback } from "react";
+
+/**
+ * Fetch the users on the current leaderboard. This is a super query
+ * that also exposes pagination.
+ */
+export const useCurrentLeaderboardUsersQuery = ({
+  initialPage = 1,
+  pageSize = 20,
+  tieToUrl = true,
+}: {
+  initialPage?: number;
+  pageSize?: number;
+  tieToUrl?: boolean;
+}) => {
+  const [page, setPage] = useURLState("page", initialPage, tieToUrl);
+
+  const goBack = useCallback(() => {
+    setPage((old) => Math.max(old - 1, 0));
+  }, [setPage]);
+
+  const goForward = useCallback(() => {
+    setPage((old) => old + 1);
+  }, [setPage]);
+
+  const goTo = useCallback(
+    (pageNumber: number) => {
+      setPage(() => Math.max(pageNumber, 0));
+    },
+    [setPage],
+  );
+
+  const query = useQuery({
+    queryKey: ["leaderboard", "users", page, pageSize],
+    queryFn: () => fetchLeaderboardUsers({ page, pageSize }),
+    placeholderData: keepPreviousData,
+  });
+
+  return {
+    ...query,
+    page,
+    goBack,
+    goForward,
+    goTo,
+  };
+};
+
+/**
+ * Fetch the details about a leaderboard (excluding users)
+ */
+export const useCurrentLeaderboardMetadataQuery = () => {
+  return useQuery({
+    queryKey: ["leaderboard", "metadata"],
+    queryFn: useCurrentLeaderboardMetadata,
+  });
+};
+
+/**
+ * Query the backend to fetch points. If it returns a successful search,
+ * we invalidate all the queries to see if anything has changed.
+ */
+export const useUsersTotalPoints = () => {
+  const queryClient = useQueryClient();
+
+  return useMutation({
+    mutationFn: updateTotalPoints,
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ["submission"] });
+      queryClient.invalidateQueries({ queryKey: ["leaderboard"] });
+      queryClient.invalidateQueries({ queryKey: ["potd"] });
+    },
+  });
+};
+
+/**
+ * To patch a rendering bug in the interim, prefetching is used to avoid any loading states.
+ */
+export const useFixMyPointsPrefetch = ({ userId }: { userId: string }) => {
+  const queryClient = useQueryClient();
+  queryClient.prefetchQuery({
+    queryKey: ["leaderboard", "user", "me"],
+    queryFn: () => getMyRecentLeaderboardData({ userId }),
+  });
+  return;
+};
+
+/**
+ * Fetch the specific submission data of the authenticated user.
+ */
+export const useMyRecentLeaderboardData = ({ userId }: { userId: string }) => {
+  return useQuery({
+    queryKey: ["leaderboard", "user", "me"],
+    queryFn: () => getMyRecentLeaderboardData({ userId }),
+  });
+};
+
+async function fetchLeaderboardUsers({
+  page,
+  pageSize,
+}: {
+  page: number;
+  pageSize: number;
+}) {
+  const response = await fetch(
+    `/api/leaderboard/current/user/all?page=${page}&pageSize=${pageSize}`,
+    {
+      method: "GET",
+    },
+  );
+
+  const json = (await response.json()) as ApiResponse<Page<UserWithScore[]>>;
+
+  return json;
+}
+
+async function useCurrentLeaderboardMetadata() {
+  const response = await fetch(`/api/leaderboard/current/metadata`, {
+    method: "GET",
+  });
+
+  const json = (await response.json()) as ApiResponse<Leaderboard>;
+
+  return json;
+}
+
+async function updateTotalPoints() {
+  const res = await fetch("/api/leetcode/check", {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+  });
+
+  const json = (await res.json()) as ApiResponse<{
+    acceptedSubmissions: { title: string; points: number }[];
+  }>;
+
+  if (res.status === 429) {
+    return { ...json, message: Number(json.message) };
+  }
+
+  return json;
+}
+
+export async function getMyRecentLeaderboardData({
+  userId,
+}: {
+  userId: string;
+}) {
+  const res = await fetch(`/api/leaderboard/current/user/${userId}`);
+
+  const json = (await res.json()) as ApiResponse<User & { totalScore: number }>;
+
+  return json;
+}
