+++ js/src/lib/api/types/leaderboard.ts
