+++ js/src/lib/api/types/potd.ts
